+++ config/sync/jsonapi_extras.settings.yml
@@ -1,5 +1,1 @@
 path_prefix: api
-dependencies:
-  module:
-    - jsonapi
-    - jsonapi_extras
