+++ modules/recipes_magazin/config/install/jsonapi_extras.jsonapi_resource_config.media_type--media_type.yml
@@ -17,109 +17,81 @@ resourceFields:
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   uuid:
     fieldName: uuid
     publicName: uuid
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   langcode:
     fieldName: langcode
     publicName: langcode
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   status:
     fieldName: status
     publicName: status
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   dependencies:
     fieldName: dependencies
     publicName: dependencies
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   third_party_settings:
     fieldName: third_party_settings
     publicName: third_party_settings
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   _core:
     fieldName: _core
     publicName: _core
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   label:
     fieldName: label
     publicName: label
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   description:
     fieldName: description
     publicName: description
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   source:
     fieldName: source
     publicName: source
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   queue_thumbnail_downloads:
     fieldName: queue_thumbnail_downloads
     publicName: queue_thumbnail_downloads
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   new_revision:
     fieldName: new_revision
     publicName: new_revision
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   source_configuration:
     fieldName: source_configuration
     publicName: source_configuration
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   field_map:
     fieldName: field_map
     publicName: field_map
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
