+++ modules/recipes_magazin/config/install/jsonapi_extras.jsonapi_resource_config.node--page.yml
@@ -118,15 +118,13 @@ resourceFields:
     enhancer:
       id: ''
   path:
-     fieldName: path
-     publicName: path
-     enhancer:
-       id: nested
-       settings:
-         path: alias
-     disabled: false
-     advancedOptionsIcon: ''
-     enhancer_label: ''
+    fieldName: path
+    publicName: path
+    enhancer:
+      id: nested
+      settings:
+        path: alias
+    disabled: false
   body:
     fieldName: body
     publicName: body
