+++ modules/recipes_magazin/config/install/jsonapi_extras.jsonapi_resource_config.node--recipe.yml
@@ -17,88 +17,72 @@ resourceFields:
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   uuid:
     disabled: true
     fieldName: uuid
     publicName: uuid
     enhancer:
       id: ''
-    advancedOptionsIcon: ''
-    enhancer_label: ''
+
   vid:
     disabled: true
     fieldName: vid
     publicName: vid
     enhancer:
       id: ''
-    advancedOptionsIcon: ''
-    enhancer_label: ''
+
   langcode:
     disabled: true
     fieldName: langcode
     publicName: langcode
     enhancer:
       id: ''
-    advancedOptionsIcon: ''
-    enhancer_label: ''
+
   type:
     fieldName: type
     publicName: contentType
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   revision_timestamp:
     disabled: true
     fieldName: revision_timestamp
     publicName: revision_timestamp
     enhancer:
       id: ''
-    advancedOptionsIcon: ''
-    enhancer_label: ''
+
   revision_uid:
     disabled: true
     fieldName: revision_uid
     publicName: revision_uid
     enhancer:
       id: ''
-    advancedOptionsIcon: ''
-    enhancer_label: ''
+
   revision_log:
     disabled: true
     fieldName: revision_log
     publicName: revision_log
     enhancer:
       id: ''
-    advancedOptionsIcon: ''
-    enhancer_label: ''
+
   status:
     fieldName: status
     publicName: isPublished
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   title:
     fieldName: title
     publicName: title
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   uid:
     fieldName: uid
     publicName: owner
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   created:
     fieldName: created
     publicName: createdAt
@@ -107,8 +91,6 @@ resourceFields:
       settings:
         dateTimeFormat: 'Y-m-d\TH:i:sO'
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   changed:
     fieldName: changed
     publicName: updatedAt
@@ -117,48 +99,39 @@ resourceFields:
       settings:
         dateTimeFormat: 'Y-m-d\TH:i:sO'
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   promote:
     fieldName: promote
     publicName: isPromoted
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   sticky:
     disabled: true
     fieldName: sticky
     publicName: sticky
     enhancer:
       id: ''
-    advancedOptionsIcon: ''
-    enhancer_label: ''
+
   default_langcode:
     disabled: true
     fieldName: default_langcode
     publicName: default_langcode
     enhancer:
       id: ''
-    advancedOptionsIcon: ''
-    enhancer_label: ''
+
   revision_default:
     fieldName: revision_default
     publicName: revision_default
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   revision_translation_affected:
     disabled: true
     fieldName: revision_translation_affected
     publicName: revision_translation_affected
     enhancer:
       id: ''
-    advancedOptionsIcon: ''
-    enhancer_label: ''
+
   path:
     fieldName: path
     publicName: path
@@ -167,93 +140,69 @@ resourceFields:
       settings:
         path: alias
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   field_author:
     fieldName: field_author
     publicName: author
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   field_cooking_time:
     fieldName: field_cooking_time
     publicName: cookingTime
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   field_difficulty:
     fieldName: field_difficulty
     publicName: difficulty
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   field_image:
     fieldName: field_image
     publicName: image
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   field_ingredients:
     fieldName: field_ingredients
     publicName: ingredients
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   field_number_of_servings:
     fieldName: field_number_of_servings
     publicName: numberOfservings
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   field_preparation_time:
     fieldName: field_preparation_time
     publicName: preparationTime
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   field_recipe_category:
     fieldName: field_recipe_category
     publicName: category
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   field_recipe_instruction:
     fieldName: field_recipe_instruction
     publicName: instructions
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   field_summary:
     fieldName: field_summary
     publicName: summary
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
   field_tags:
     fieldName: field_tags
     publicName: tags
     enhancer:
       id: ''
     disabled: false
-    advancedOptionsIcon: ''
-    enhancer_label: ''
