+++ modules/recipes_magazin/config/install/jsonapi_extras.jsonapi_resource_config.node--tutorial.yml
@@ -116,15 +116,13 @@ resourceFields:
     enhancer:
       id: ''
   path:
-     fieldName: path
-     publicName: path
-     enhancer:
-       id: nested
-       settings:
-         path: alias
-     disabled: false
-     advancedOptionsIcon: ''
-     enhancer_label: ''
+    fieldName: path
+    publicName: path
+    enhancer:
+      id: nested
+      settings:
+        path: alias
+    disabled: false
   field_image:
     fieldName: field_image
     publicName: image
