+++ modules/recipes_magazin/config/install/jsonapi_extras.jsonapi_resource_config.responsive_image_style--responsive_image_style.yml
@@ -10,4 +10,4 @@ id: responsive_image_style--responsive_image_style
 disabled: true
 path: responsive_image_style/responsive_image_style
 resourceType: responsive_image_style--responsive_image_style
-resourceFields: ''
+resourceFields: []
