+++ modules/recipes_magazin/config/install/jsonapi_extras.jsonapi_resource_config.taxonomy_term--category.yml
@@ -76,10 +76,8 @@ resourceFields:
     enhancer:
       id: ''
   path:
-     fieldName: path
-     publicName: path
-     enhancer:
-       id: ''
-     disabled: false
-     advancedOptionsIcon: ''
-     enhancer_label: ''
+    fieldName: path
+    publicName: path
+    enhancer:
+      id: ''
+    disabled: false
