+++ modules/recipes_magazin/config/install/views.view.oauth_tutorials.yml
